docs(site): add Code tab to use-position demo (core + harness)
Extract clean UsePositionExample (side/align props, self-managed open state)
from UsePositionDemo; rewrite UsePositionDemo as a placement-explorer harness
with fieldset controls that drive the core. Wire use-position.mdx to show
UsePositionExample via the Code tab and drop the redundant Example fence.
