+++ apps/site/src/components/docs/UsePositionDemo.tsx
@@ -1,23 +1,15 @@
-import { usePosition } from 'hono-preact-ui';
-import { useRef, useState } from 'preact/hooks';
+import { useState } from 'preact/hooks';
+import { UsePositionExample } from './UsePositionExample.js';
 
 const SIDES = ['top', 'right', 'bottom', 'left'] as const;
 const ALIGNS = ['start', 'center', 'end'] as const;
 
-// Bare placement: usePosition anchors the floating box to the button and applies
-// position:fixed/left/top itself, resolving a final side/align after collision
-// handling (the readout shows the resolved values, which can differ from the
-// requested side near a viewport edge). Deliberately not a popover. Styling:
-// .docs-useposition* in root.css.
+// Placement explorer harness around the UsePositionExample core. The fieldset
+// controls exist only here (the docs Code tab shows UsePositionExample, the real
+// usage). Styling: .docs-useposition* in root.css.
 export function UsePositionDemo() {
-  const anchorRef = useRef<HTMLButtonElement>(null);
-  const floatingRef = useRef<HTMLDivElement>(null);
-  const [open, setOpen] = useState(false);
   const [side, setSide] = useState<(typeof SIDES)[number]>('bottom');
   const [align, setAlign] = useState<(typeof ALIGNS)[number]>('center');
-
-  const pos = usePosition({ open, anchorRef, floatingRef, side, align });
-
   return (
     <div class="docs-useposition">
       <div class="docs-useposition-controls">
@@ -50,24 +42,7 @@ export function UsePositionDemo() {
           ))}
         </fieldset>
       </div>
-      <button
-        ref={anchorRef}
-        type="button"
-        class="docs-useposition-anchor"
-        onClick={() => setOpen((o) => !o)}
-      >
-        {open ? 'Hide' : 'Show'} box
-      </button>
-      {open ? (
-        <div
-          ref={floatingRef}
-          class="docs-useposition-box"
-          data-side={pos.side}
-          data-align={pos.align}
-        >
-          resolved: {pos.side} / {pos.align}
-        </div>
-      ) : null}
+      <UsePositionExample side={side} align={align} />
     </div>
   );
 }
